// packages/react/src/components/Button/Enum.ts
export enum ButtonVariant {
  PRIMARY = 'primary',
  PRIMARY_OUTLINE = 'primary_outline',
  SECONDARY = 'secondary',
  ACCENT = 'accent',
  ACCENT_OUTLINE = 'accent_outline',
  POSITIVE = 'positive',
  POSITIVE_OUTLINE = 'positive_outline',
  NEGATIVE = 'negative',
  NEGATIVE_OUTLINE = 'negative_outline',
  WHITE = 'white',
  WHITE_OUTLINE = 'white_outline',
  LINK = 'link',
}

export enum ButtonSize {
  SMALL = 'small',
  MEDIUM = 'medium',
  LARGE = 'large',
}

export enum ButtonWidth {
  FULL = 'full',
  AUTO = 'auto',
  MIN = 'min',
}